Update README.md
The verification system is a Devices Status Verification web-based software for verifying the status of devices. While buying the stolen product which is to be sold at the shop and to people will be verified online. It will show the status of the product if it is stolen, Lost, or cleared. Three main actors interact with the website i.e. Main Admin, Station, and User. The main admin can update categories and stations which can upload products and it also adds products. The user will verify the product with or with making account. If the user makes an account on the website it will keep his history of searching products, which will be cleared manually in the dashboard.